exec: Put provider in a subprocess (redesign)
This change makes mairu-exec to spawn a credential provider in a child
process, instead of: run a target command in a child process which were
formerly done. We now call a child process running a credential provider
'sidecar'. A sidecar will be launched via fork(2).

And a parent process; an original mairu-exec process ran by user is
now responsible to execvp(2) a specified target command. We now call the
parent process an 'executor'.

This strategy lets Mairu transparent from other processes, in terms of
handling signal and returning exit status; including signals (WTERMSIG).
OTOH, other processes that execute mairu-exec doesn't have to care about
a weird intermediate process (mairu-exec) anymore. Previously an user
has to take care of, for instance specifying --forward-signals.

Now mairu-exec spawns a sidecar by fork(2) with pipe(2) and the mairu-exec
itself (parent process) calls an executor function. The spawned sidecar
prepares a credential provider as like as we've done before, and informs
the executor when it become ready. The executor waits for an information
from the sidecar, then execvp(2) a specified target command line with
applying the given information, for instance, AWS_* environment variables.

The sidecar monitors the parent process (formerly an executor) while it
keeps provider running, and it quits when the parent is gone.

There is an optimization path to notice a parent dismissal on Linux,
FreeBSD, DragonFlyBSD thru signal, but at this moment a successful
build is only verified on Linux.

This design is inspired by ssh-agent(1); a ssh-agent will run in a child
process of a given command of ssh-agent(1) and terminates when the
command ends.